Fix hook patch ABI helpers and defer render-driver startup

+++ rw_rh_engine_lib/render_client/render_client.cpp
@@ -36,26 +36,58 @@ RenderClient::RenderClient()
             .mSize = 1024 * 1024 *
                      rh::engine::EngineConfigBlock::It.SharedMemorySizeMB,
             .mOwner = true } );
-
-    /// Create render driver sub-process
-    STARTUPINFOA start_info{ .cb = sizeof( start_info ) };
-    CreateProcessA( IPCSettings::mProcessName.c_str(), nullptr, nullptr,
-                    nullptr, false, 0, nullptr, nullptr, &start_info,
-                    &RenderDriverProcess );
 }
 
 RenderClient::~RenderClient()
 {
     TaskQueue->SendExitEvent();
     TaskQueue.reset();
-    if ( RenderDriverProcess.hProcess )
+    if ( RenderDriverStarted && RenderDriverProcess.hProcess )
         TerminateProcess( RenderDriverProcess.hProcess, 0 );
+    if ( RenderDriverStarted && RenderDriverProcess.hThread )
+        CloseHandle( RenderDriverProcess.hThread );
+    if ( RenderDriverStarted && RenderDriverProcess.hProcess )
+        CloseHandle( RenderDriverProcess.hProcess );
 }
 
 bool RenderClient::RegisterPlugins( const PluginPtrTable &plugin_cb )
 {
     Plugins = std::make_unique<ClientPlugins>( plugin_cb );
     return true;
+}
+
+bool RenderClient::StartRenderDriverProcess()
+{
+    if ( RenderDriverStarted )
+        return true;
+
+    if ( IPCSettings::mProcessName.empty() )
+    {
+        debug::DebugLogger::Log(
+            "Render driver process name is empty.",
+            debug::LogLevel::Error );
+        return false;
+    }
+
+    STARTUPINFOA start_info{ .cb = sizeof( start_info ) };
+    if ( !CreateProcessA( IPCSettings::mProcessName.c_str(), nullptr, nullptr,
+                          nullptr, false, 0, nullptr, nullptr, &start_info,
+                          &RenderDriverProcess ) )
+    {
+        debug::DebugLogger::Log(
+            "Failed to start render driver process: " +
+                IPCSettings::mProcessName,
+            debug::LogLevel::Error );
+        return false;
+    }
+
+    RenderDriverStarted = true;
+    return true;
+}
+
+void RenderClient::EnsureRenderDriverStarted()
+{
+    StartRenderDriverProcess();
 }
 
 } // namespace rh::rw::engine
